+++ pybamm/models/submodels/thermal/pouch_cell/__init__.py
@@ -1,0 +1,2 @@
+from .pouch_cell_1D_current_collectors import CurrentCollector1D
+from .pouch_cell_2D_current_collectors import CurrentCollector2D