+++ tests/unit/test_models/test_full_battery_models/test_lithium_ion/test_dfn.py
@@ -38,15 +38,15 @@ class TestDFN(unittest.TestCase):
         with self.assertRaises(pybamm.OptionError):
             model = pybamm.lithium_ion.DFN(options)
 
-    def test_x_full_thermal_model_no_current_collector(self):
-        options = {"thermal": "x-full"}
+    def test_lumped_thermal_model_1D(self):
+        options = {"thermal": "x-lumped"}
         model = pybamm.lithium_ion.DFN(options)
         model.check_well_posedness()
 
-        # Not implemented with current collectors
-        options = {"thermal": "x-full", "thermal current collector": True}
-        with self.assertRaises(NotImplementedError):
-            model = pybamm.lithium_ion.DFN(options)
+    def test_x_full_thermal_model(self):
+        options = {"thermal": "x-full"}
+        model = pybamm.lithium_ion.DFN(options)
+        model.check_well_posedness()
 
     def test_x_full_Nplus1D_not_implemented(self):
         # 1plus1D
@@ -66,39 +66,7 @@ class TestDFN(unittest.TestCase):
         with self.assertRaises(NotImplementedError):
             pybamm.lithium_ion.DFN(options)
 
-    def test_x_lumped_thermal_model_no_Current_collector(self):
-        options = {"thermal": "x-lumped"}
-        model = pybamm.lithium_ion.DFN(options)
-        model.check_well_posedness()
-
-        # xyz-lumped returns the same as x-lumped
-        options = {"thermal": "xyz-lumped"}
-        model = pybamm.lithium_ion.DFN(options)
-        model.check_well_posedness()
-
-    def test_x_lumped_thermal_model_0D_current_collector(self):
-        options = {"thermal": "x-lumped", "thermal current collector": True}
-        model = pybamm.lithium_ion.DFN(options)
-        model.check_well_posedness()
-
-        # xyz-lumped returns the same as x-lumped
-        options = {"thermal": "xyz-lumped", "thermal current collector": True}
-        model = pybamm.lithium_ion.DFN(options)
-        model.check_well_posedness()
-
-        options = {"thermal": "lumped"}
-        model = pybamm.lithium_ion.DFN(options)
-        model.check_well_posedness()
-
-    def test_xyz_lumped_thermal_1D_current_collector(self):
-        options = {
-            "current collector": "potential pair",
-            "dimensionality": 1,
-            "thermal": "xyz-lumped",
-        }
-        model = pybamm.lithium_ion.DFN(options)
-        model.check_well_posedness()
-
+    def test_lumped_thermal_1plus1D(self):
         options = {
             "current collector": "potential pair",
             "dimensionality": 1,
@@ -107,15 +75,7 @@ class TestDFN(unittest.TestCase):
         model = pybamm.lithium_ion.DFN(options)
         model.check_well_posedness()
 
-    def test_xyz_lumped_thermal_2D_current_collector(self):
-        options = {
-            "current collector": "potential pair",
-            "dimensionality": 2,
-            "thermal": "xyz-lumped",
-        }
-        model = pybamm.lithium_ion.DFN(options)
-        model.check_well_posedness()
-
+    def test_lumped_thermal_2plus1D(self):
         options = {
             "current collector": "potential pair",
             "dimensionality": 2,
@@ -124,7 +84,7 @@ class TestDFN(unittest.TestCase):
         model = pybamm.lithium_ion.DFN(options)
         model.check_well_posedness()
 
-    def test_x_lumped_thermal_1D_current_collector(self):
+    def test_thermal_1plus1D(self):
         options = {
             "current collector": "potential pair",
             "dimensionality": 1,
@@ -133,7 +93,7 @@ class TestDFN(unittest.TestCase):
         model = pybamm.lithium_ion.DFN(options)
         model.check_well_posedness()
 
-    def test_x_lumped_thermal_2D_current_collector(self):
+    def test_thermal_2plus1D(self):
         options = {
             "current collector": "potential pair",
             "dimensionality": 2,