+++ tests/unit/test_models/test_submodels/test_thermal/test_lumped.py
@@ -1,5 +1,5 @@
 #
-# Test xyz-lumped submodel with 2D current collectors
+# Test lumped thermal submodel
 #
 
 import pybamm
@@ -11,20 +11,19 @@ from tests.unit.test_models.test_submodels.test_thermal.coupled_variables import
 )
 
 
-class TestCurrentCollector2D(unittest.TestCase):
+class TestLumped(unittest.TestCase):
     def test_public_functions(self):
         param = pybamm.standard_parameters_lithium_ion
-        phi_s_cn = pybamm.PrimaryBroadcast(pybamm.Scalar(0), ["current collector"])
-        phi_s_cp = pybamm.PrimaryBroadcast(pybamm.Scalar(3), ["current collector"])
 
-        coupled_variables.update(
-            {
-                "Negative current collector potential": phi_s_cn,
-                "Positive current collector potential": phi_s_cp,
-            }
-        )
+        submodel = pybamm.thermal.Lumped(param)
+        std_tests = tests.StandardSubModelTests(submodel, coupled_variables)
+        std_tests.test_all()
 
-        submodel = pybamm.thermal.xyz_lumped.CurrentCollector2D(param)
+        submodel = pybamm.thermal.Lumped(param, cc_dimension=1)
+        std_tests = tests.StandardSubModelTests(submodel, coupled_variables)
+        std_tests.test_all()
+
+        submodel = pybamm.thermal.Lumped(param, cc_dimension=2)
         std_tests = tests.StandardSubModelTests(submodel, coupled_variables)
         std_tests.test_all()