+++ tests/unit/test_models/test_submodels/test_thermal/test_pouch_cell/test_pouch_2plus1D.py
@@ -1,5 +1,5 @@
 #
-# Test x-lumped submodel with 2D current collectors
+# Test 2+1D submodel
 #
 
 import pybamm
@@ -11,20 +11,10 @@ from tests.unit.test_models.test_submodels.test_thermal.coupled_variables import
 )
 
 
-class TestCurrentCollector2D(unittest.TestCase):
+class TestPouchCell2D(unittest.TestCase):
     def test_public_functions(self):
         param = pybamm.standard_parameters_lithium_ion
-        phi_s_cn = pybamm.PrimaryBroadcast(pybamm.Scalar(0), ["current collector"])
-        phi_s_cp = pybamm.PrimaryBroadcast(pybamm.Scalar(3), ["current collector"])
-
-        coupled_variables.update(
-            {
-                "Negative current collector potential": phi_s_cn,
-                "Positive current collector potential": phi_s_cp,
-            }
-        )
-
-        submodel = pybamm.thermal.x_lumped.CurrentCollector2D(param)
+        submodel = pybamm.thermal.pouch_cell.CurrentCollector2D(param)
         std_tests = tests.StandardSubModelTests(submodel, coupled_variables)
         std_tests.test_all()