+++ tests/unit/test_simulation.py
@@ -193,7 +193,7 @@ class TestSimulation(unittest.TestCase):
         self.assertIsInstance(c_e, np.ndarray)
 
     def test_set_external_variable(self):
-        model_options = {"thermal": "x-lumped", "external submodels": ["thermal"]}
+        model_options = {"thermal": "lumped", "external submodels": ["thermal"]}
         model = pybamm.lithium_ion.SPMe(model_options)
         sim = pybamm.Simulation(model)
 
@@ -201,7 +201,7 @@ class TestSimulation(unittest.TestCase):
 
         dt = 0.001
 
-        external_variables = {"X-averaged cell temperature": T_av}
+        external_variables = {"Volume-averaged cell temperature": T_av}
         sim.step(dt, external_variables=external_variables)
 
     def test_step(self):